requirements.txt
# training requirements
vllm
xformers
transformers
torch
accelerate
deepspeed
peft
bitsandbytes
flash-attn
# llm
openai
anthropic
# auxilary libary
trl
sentence-transformers
sentencepiece
scikit-learn
datasets
evaluate
notebook
matplotlib
pandas
wandb
gradio
colorlog
hydra-core
einops
tensorboard
unsloth